Directories organized

+++ app/src/main/java/com/example/dictionaryapp/adapters/RVAdapterHistoryFragment.kt
@@ -1,26 +1,22 @@
-package com.example.dictionaryapp.view.adapters
+package com.example.dictionaryapp.adapters
 
 import android.annotation.SuppressLint
 import android.content.Context
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
-import android.widget.Toast
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import com.example.dictionaryapp.R
-import com.example.dictionaryapp.view.db_favorites.DatabaseHelperFavorites
-import com.example.dictionaryapp.view.db_favorites.Favorites
-import com.example.dictionaryapp.view.db_favorites.FavoritesDao
-import com.example.dictionaryapp.view.db_history.DatabaseHelper
-import com.example.dictionaryapp.view.db_history.Histories
-import com.example.dictionaryapp.view.db_history.HistoriesDao
+import com.example.dictionaryapp.db.db_favorites.DatabaseHelperFavorites
+import com.example.dictionaryapp.db.db_favorites.FavoritesDao
+import com.example.dictionaryapp.db.db_history.DatabaseHelper
+import com.example.dictionaryapp.db.db_history.Histories
+import com.example.dictionaryapp.db.db_history.HistoriesDao
 
-class RVAdapterFavoritesFragment (private val mContext: Context, private var historyList:List<Favorites>, val listener: (String) -> Unit) : RecyclerView.Adapter<RVAdapterFavoritesFragment.CardDesignObjectsHolder>() {
+class RVAdapterHistoryFragment (private val mContext: Context, private var historyList:List<Histories>, val listener: (String) -> Unit) : RecyclerView.Adapter<RVAdapterHistoryFragment.CardDesignObjectsHolder>() {
 
     private lateinit var dbh: DatabaseHelper
     private lateinit var dbhf: DatabaseHelperFavorites
@@ -53,7 +49,9 @@ class RVAdapterFavoritesFragment (private val mContext: Context, private var his
         dbh = DatabaseHelper(mContext)
         dbhf = DatabaseHelperFavorites(mContext)
 
-        holder.imageViewBookmark.setImageResource(R.drawable.ic_baseline_bookmark_24)
+        if (FavoritesDao().isInFavorite(dbhf, history.word) == 1) {
+            holder.imageViewBookmark.setImageResource(R.drawable.ic_baseline_bookmark_24)
+        }
 
         holder.textViewWord.text = history.word
 
@@ -71,17 +69,18 @@ class RVAdapterFavoritesFragment (private val mContext: Context, private var his
          */
 
         holder.imageViewBookmark.setOnClickListener {
-            HistoriesDao().removeFavorites(dbh, history.word)
-            FavoritesDao().deleteFavorites(dbhf, history.word)
-            holder.imageViewBookmark.setImageResource(R.drawable.ic_bookmark)
-            notifyItemRemoved(position)
-            historyList = FavoritesDao().getFavorites(dbhf)
-            notifyItemRangeChanged(position, historyList.size)
-            Log.e("list", historyList.toString())
+            if (FavoritesDao().isInFavorite(dbhf, history.word) == 1) {
+                HistoriesDao().removeFavorites(dbh, history.word)
+                FavoritesDao().deleteFavorites(dbhf, history.word)
+                holder.imageViewBookmark.setImageResource(R.drawable.ic_bookmark)
+                historyList = HistoriesDao().getFavorites(dbh)
+            } else {
+                holder.imageViewBookmark.setImageResource(R.drawable.ic_baseline_bookmark_24)
+                HistoriesDao().addToFavorites(dbh, history.word)
+                FavoritesDao().addFavorites(dbhf, dbh, history.word)
+            }
             // HistoriesDao().addToFavorites(dbh, history.word)
             // Toast.makeText(mContext, "${history.word} bookmark clicked", Toast.LENGTH_SHORT).show()
-            it.findNavController().popBackStack(R.id.favoritesFragment,true)
-            it.findNavController().navigate(R.id.favoritesFragment)
         }
 
         holder.linearLayoutDesign.setOnClickListener { listener(history.word) }
